Feat/b/status code

See merge request s08-webmobile1-sub2/S08P12A606!29

+++ backend/Homezakaya/src/main/java/com/ssafy/Homezakaya/controller/UserInRoomController.java
@@ -1,12 +1,18 @@
 package com.ssafy.Homezakaya.controller;
 
 import com.ssafy.Homezakaya.model.dto.UserInRoomDto;
+import com.ssafy.Homezakaya.model.service.RoomServiceImpl;
 import com.ssafy.Homezakaya.model.service.UserInRoomServiceImpl;
+import com.ssafy.Homezakaya.model.service.UserService;
+import com.ssafy.Homezakaya.model.service.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
+import java.util.HashMap;
+import java.util.Map;
 
 @RequestMapping("/api/userinroom")
 @RestController
@@ -15,8 +21,32 @@ public class UserInRoomController {
     @Autowired
     private UserInRoomServiceImpl userInRoomService;
 
+    @Autowired
+    private UserServiceImpl userService;
+    @Autowired
+    private RoomServiceImpl roomService;
+
     @PostMapping
     public ResponseEntity<?> createUserInRoom(@RequestBody UserInRoomDto userInRoom) {
+        Map<String, Object> resultMap = new HashMap<>();
+        // 아이디가 존재하지 않음
+        if(userService.getUser(userInRoom.getUserId()) == null){
+            resultMap.put("message", "존재하지 않는 userId");
+            return new ResponseEntity<Map<String, Object>>(resultMap, HttpStatus.NOT_FOUND);
+        }
+
+        // 해당 roomId가 존재하지 않음
+        if(roomService.getRoom(userInRoom.getRoomId()) == null){
+            resultMap.put("message", "존재하지 않는 roomId");
+            return new ResponseEntity<Map<String, Object>>(resultMap, HttpStatus.NOT_FOUND);
+        }
+
+        // 해당 유저가 이미 방에 참가한 상태
+        if(userInRoomService.getRoomId(userInRoom.getUserId()) != null){
+            resultMap.put("message", "이미 해당 room에 참가한 유저입니다.");
+            return new ResponseEntity<Map<String, Object>>(resultMap, HttpStatus.CONFLICT);
+        }
+
         boolean res = userInRoomService.createUserInRoom(userInRoom);
 
         if(res)
@@ -27,6 +57,19 @@ public class UserInRoomController {
 
     @DeleteMapping("/{userId}")
     public ResponseEntity<?> removeUserInRoom(@PathVariable String userId) {
+        Map<String, Object> resultMap = new HashMap<>();
+        // 아이디가 존재하지 않음
+        if(userService.getUser(userId) == null){
+            resultMap.put("message", "존재하지 않는 userId");
+            return new ResponseEntity<Map<String, Object>>(resultMap, HttpStatus.NOT_FOUND);
+        }
+
+        // 유저가 room에 참가한 상태가 아님
+        if(userInRoomService.getRoomId(userId) == null){
+            resultMap.put("message", "room에 참가한 유저가 아닙니다.");
+            return new ResponseEntity<Map<String, Object>>(resultMap, HttpStatus.NOT_FOUND);
+        }
+
         boolean res = userInRoomService.removeUserInRoom(userId);
 
         if(res)
@@ -37,8 +80,21 @@ public class UserInRoomController {
 
     @GetMapping("/{userId}")
     public ResponseEntity<?> getRoomId(@PathVariable String userId){
+        Map<String, Object> resultMap = new HashMap<>();
+        // 아이디가 존재하지 않음
+        if(userService.getUser(userId) == null){
+            resultMap.put("message", "존재하지 않는 userId");
+            return new ResponseEntity<Map<String, Object>>(resultMap, HttpStatus.NOT_FOUND);
+        }
+
+        // 유저가 room에 참가한 상태가 아님
+        if(userInRoomService.getRoomId(userId) == null){
+            resultMap.put("message", "room에 참가한 유저가 아닙니다.");
+            return new ResponseEntity<Map<String, Object>>(resultMap, HttpStatus.NOT_FOUND);
+        }
+
         Integer res = userInRoomService.getRoomId(userId);
-        System.out.println(res);
+
         if(res != null)
             return ResponseEntity.ok(res);
         else
